main_api/requirements.txt
fastapi==0.119.0
psycopg2-binary==2.9.11
elasticsearch==9.1.1
uvicorn==0.38.0
aiohttp==3.13.1
redis==7.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
jinja2==3.1.6
backoff==2.2.1
pyjwt==2.10.1